ECA-10132 Use dummy pin for autoactiv. when Azure uses key binding.
Azure Key Vault Crypto Tokens set a dummy pin when using key bindings to
authenticate.  This is because EJBCA uses the existence of an
autoactivate ping as the indicator that a token auto-activates.
Obviously, there is no PIN for Azure Tokens using key bindings to
authenticate.

Also, I hit "reformat" while editing the code, so there are a bunch of
whitespace changes.